+++ Sources/DockSDK/SDKVersion.swift
@@ -6,8 +6,8 @@ import Foundation
 
 @objc(DockSDKVersion)
 public class DockSDKVersion: NSObject {
-    @objc public static let current: String = "1.0.0"
+    @objc public static let current: String = "1.2.0"
     @objc public static let major: Int = 1
-    @objc public static let minor: Int = 0
+    @objc public static let minor: Int = 2
     @objc public static let patch: Int = 0
 }